改行，export等の簡単なリファクタリング
