+++ src/components/Footer.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {Link} from "react-router-dom"
 
-const Footer = () => {
+export const Footer = () => {
   return (
     <footer>
       <nav>
@@ -10,9 +10,6 @@ const Footer = () => {
             <li>
               <Link to="/">ホーム</Link>
             </li>
-            {/* <li>
-              <Link to="/about">about</Link>
-            </li> */}
           </ul>
           <p className="copyright">© 2023 ganmolt</p>
         </div>
@@ -20,5 +17,3 @@ const Footer = () => {
     </footer>
   );
 };
-
-export default Footer;
