+++ src/components/Header.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {Link} from "react-router-dom"
 
-const Header = () => {
+export const Header = () => {
   return <header>
     <div className="logo">
       <h3>ポートフォリオ</h3>
@@ -12,12 +12,7 @@ const Header = () => {
         <li>
           <Link to="/">ホーム</Link>
         </li>
-        {/* <li>
-          <Link to="/about">about</Link>
-        </li> */}
       </ul>
     </nav>
   </header>;
 };
-
-export default Header;
