+++ src/index.js
@@ -10,4 +10,3 @@ root.render(
     <App />
   </React.StrictMode>
 );
-
